a4(single node functionality untested

+++ src/server/Constants/serverConstants.go
@@ -5,7 +5,7 @@ import (
 	"os"
 )
 
-var PID = os.Getpid()
+var PID = int32(os.Getpid())
 
 const (
 	// Server constants
